+++ src/compiler/types/tests/generate-prop-types.spec.ts
@@ -20,10 +20,9 @@ describe('generate-prop-types', () => {
       updateTypeIdentifierNamesSpy.mockImplementation(
         (
           _typeReferences: d.ComponentCompilerTypeReferences,
-          _cmpMeta: d.ComponentCompilerMeta,
           _typeImportData: d.TypesImportData,
-          initialType: string,
-          _updateTypeName: (currentTypeName: string, typeAlias: d.TypesMemberNameData) => string
+          _sourceFilePath: string,
+          initialType: string
         ) => initialType
       );