+++ src/declarations/stencil-public-compiler.ts
@@ -278,6 +278,14 @@ export interface ConfigExtras {
   dynamicImportShim?: boolean;
 
   /**
+   * Experimental flag. Projects that use a Stencil library built using the `dist` output target may have trouble lazily
+   * loading components when using a bundler such as Vite or Parcel. Setting this flag to `true` will change how Stencil
+   * lazily loads components in a way that works with additional bundlers. Setting this flag to `true` will increase
+   * the size of the compiled output. Defaults to `false`.
+   */
+  experimentalImportInjection?: boolean;
+
+  /**
    * Dispatches component lifecycle events. Mainly used for testing. Defaults to `false`.
    */
   lifecycleDOMEvents?: boolean;
@@ -351,6 +359,40 @@ export interface Config extends StencilConfig {
   _isValidated?: boolean;
   _isTesting?: boolean;
 }
+
+/**
+ * A 'loose' type useful for wrapping an incomplete / possible malformed
+ * object as we work on getting it comply with a particular Interface T.
+ *
+ * Example:
+ *
+ * ```ts
+ * interface Foo {
+ *   bar: string
+ * }
+ *
+ * function validateFoo(foo: Loose<Foo>): Foo {
+ *   let validatedFoo = {
+ *     ...foo,
+ *     bar: foo.bar || DEFAULT_BAR
+ *   }
+ *
+ *   return validatedFoo
+ * }
+ * ```
+ *
+ * Use this when you need to take user input or something from some other part
+ * of the world that we don't control and transform it into something
+ * conforming to a given interface. For best results, pair with a validation
+ * function as shown in the example.
+ */
+type Loose<T extends Object> = Record<string, any> & Partial<T>;
+
+/**
+ * A Loose version of the Config interface. This is intended to let us load a partial config
+ * and have type information carry though as we construct an object which is a valid `Config`.
+ */
+export type UnvalidatedConfig = Loose<Config>;
 
 export interface HydratedFlag {
   /**
@@ -1876,9 +1918,23 @@ export interface OutputTargetCustom extends OutputTargetBase {
   copy?: CopyTask[];
 }
 
+/**
+ * Output target for generating [custom data](https://github.com/microsoft/vscode-custom-data) for VS Code as a JSON
+ * file.
+ */
 export interface OutputTargetDocsVscode extends OutputTargetBase {
+  /**
+   * Designates this output target to be used for generating VS Code custom data.
+   * @see OutputTargetBase#type
+   */
   type: 'docs-vscode';
+  /**
+   * The location on disk to write the JSON file.
+   */
   file: string;
+  /**
+   * A base URL to find the source code of the component(s) described in the JSON file.
+   */
   sourceCodeBaseUrl?: string;
 }
 
@@ -1946,7 +2002,13 @@ export interface OutputTargetDistCustomElementsBundle extends OutputTargetBaseNe
   minify?: boolean;
 }
 
+/**
+ * The base type for output targets. All output targets should extend this base type.
+ */
 export interface OutputTargetBase {
+  /**
+   * A unique string to differentiate one output target from another
+   */
   type: string;
 }
 
@@ -2104,7 +2166,7 @@ export interface LoadConfigInit {
    * User config object to merge into default config and
    * config loaded from a file path.
    */
-  config?: Config;
+  config?: UnvalidatedConfig;
   /**
    * Absolute path to a Stencil config file. This path cannot be
    * relative and it does not resolve config files within a directory.
@@ -2120,8 +2182,13 @@ export interface LoadConfigInit {
   initTsConfig?: boolean;
 }
 
+/**
+ * Results from an attempt to load a config. The values on this interface
+ * have not yet been validated and are not ready to be used for arbitrary
+ * operations around the codebase.
+ */
 export interface LoadConfigResults {
-  config: Config;
+  config: UnvalidatedConfig;
   diagnostics: Diagnostic[];
   tsconfig: {
     path: string;